requirements.txt
Django
psycopg2
gunicorn
celery
redis
django-tailwind[reload]
cookiecutter

djangorestframework==3.14.0
drf-spectacular

wagtail
Pillow
openai

django-salesman
wagtail-modeladmin
social-auth-app-django
django-meta

wagtail-autocomplete
wagtail-icon-chooser
wagtail-heroicons
wagtailvideos

python-slugify
pandas

vtracer
